Update README.md
Added a sidebar to the Wiki page documents. Corrected the Wiki Home URL in the README.md file. Now that I've done this multiple times, I've realized that I didn't need to put numbers on each page to keep them in order. Lesson learned how to properly set the order of pages and will be correct on future projects.